UPSTREAM: binderfs: rename header to binderfs.h
It doesn't make sense to call the header binder_ctl.h when its sole
existence is tied to binderfs. So give it a sensible name. Users will far
more easily remember binderfs.h than binder_ctl.h.


Bug: 136497735
(cherry picked from commit c13295ad219d8bb0e47942d4cfc8251de449a67e)
